Rename active_record_patches to active_record_5_2
