+++ lib/paranoia/active_record_5_2.rb
